clone 10 times max

+++ .ci/clone.sh
@@ -1,5 +1,13 @@
+TRIES=0
+
 while ! git clone $FORGEJO_CLONE_URL eden; do
   echo "Clone failed!"
+  TRIES=$(($TRIES + 1))
+  if [ "$TRIES" = 10 ]; then
+    echo "Failed to clone after ten tries. Exiting."
+    exit 1
+  fi
+
   sleep 5
   echo "Trying clone again..."
   rm -rf ./eden || true
